Conflicts:
	GUI/Component/GameObject.java
	bin/Component/GameObject.class

+++ GUI/Component/Wall.java
@@ -4,7 +4,7 @@ import java.awt.geom.Point2D;
 
 public class Wall extends GameObject{
   
-  public Wall(Point2D.Double loc, int dir, Maze aMap)
+  public Wall(Point2D.Double loc, double dir, Maze aMap)
   {
     super(loc, dir, aMap);
   }
